% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sequences.R
\name{.BioTokenizeR_annotate_RNA}
\alias{.BioTokenizeR_annotate_RNA}
\title{Annotate RNA Sequences for BPE Workflows}
\usage{
.BioTokenizeR_annotate_RNA(seqs)
}
\arguments{
\item{seqs}{A \code{Biostrings::RNAStringSet} object containing preprocessed DNA
sequences.}
}
\value{
A list with components:
\describe{
\item{\code{seqs}}{The input RNA sequences updated with metadata columns
(\code{length} and \code{gc_content}).}
\item{\code{steps}}{A character vector of annotation steps applied.}
}
}
\description{
Computes basic annotations for DNA sequences, including sequence length and
GC content, for downstream byte-pair encoding (BPE) analysis.
}
\seealso{
Other annotation: 
\code{\link{.BioTokenizeR_annotate_AA}()},
\code{\link{.BioTokenizeR_annotate_DNA}()},
\code{\link{annotate_sequences}()}
}
\concept{annotation}
\keyword{annotation}
\keyword{internal}
