% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_scoring.R
\name{.BioTokenizeR_compute_bio_score}
\alias{.BioTokenizeR_compute_bio_score}
\title{Compute Biological Score for BPE Tokenization}
\usage{
.BioTokenizeR_compute_bio_score(bioBPE_seqs)
}
\arguments{
\item{bioBPE_seqs}{A \code{bioBPE_preprocessed} object containing preprocessed
DNA, RNA, or AA sequences.}
}
\value{
A numeric vector of biology scores per sequence sorted in the same
order as \code{bioBPE_seqs}.
}
\description{
Computes a biological score from the provided annotations as input into
biology-aware BPE tokenization.
}
\seealso{
Other preprocessing: 
\code{\link{.BioTokenizeR_preprocess_AA}()},
\code{\link{.BioTokenizeR_preprocess_DNA}()},
\code{\link{.BioTokenizeR_preprocess_RNA}()},
\code{\link{preprocess_sequences}()}
}
\concept{preprocessing}
\keyword{internal}
\keyword{preprocessing}
