% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_sequences.R
\name{.BioTokenizeR_merge_best_pair}
\alias{.BioTokenizeR_merge_best_pair}
\title{Merge a Specified Token Pair in Integer-Encoded Sequences}
\usage{
.BioTokenizeR_merge_best_pair(id_seqs, a, b, new_id)
}
\arguments{
\item{id_seqs}{A list of integer vectors representing tokenized sequences.}

\item{a}{Integer ID of the first token in the pair to merge.}

\item{b}{Integer ID of the second token in the pair to merge.}

\item{new_id}{Integer ID to assign to the merged token.}
}
\value{
A list of integer vectors, with the specified pairs replaced by
\code{new_id}.
}
\description{
Replaces all occurrences of a specified adjacent token pair \verb{(a, b)} in a
list of integer-encoded sequences with a new merged token ID. Sequences
that are too short or do not contain the pair are returned unchanged.
}
\references{
{
Medvedev A, Viswanathan K, Kanithi P (2025).BioToken and BioFM -
Biologically‑Informed Tokenization Framework. bioRxiv.
https://doi.org/10.1101/2025.03.27.645711
}
}
\seealso{
Other tokenization: 
\code{\link{.BioTokenizeR_apply_bpe}()},
\code{\link{.BioTokenizeR_compute_bio_pair_frequencies}()},
\code{\link{.BioTokenizeR_learn_bpe_vocabulary}()},
\code{\link{tokenize_sequences}()}
}
\concept{tokenization}
\keyword{internal}
\keyword{tokenization}
