% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_scoring.R
\name{.BioTokenizeR_normalize}
\alias{.BioTokenizeR_normalize}
\title{Normalize Biological Annotation}
\usage{
.BioTokenizeR_normalize(x)
}
\arguments{
\item{x}{A numerical value to min-max scale.}
}
\value{
A numerical value scaled to the range of (0, 1).
}
\description{
This function normalizes a biological annotation of a DNA, RNA, or AA
sequence through min-max scaling to (0, 1). Scaling is necessary to ensure
comparability between sequences within a set.
}
\references{
{
Wickham H, Pedersen T, Seidel D (2025). \emph{scales: Scale Functions for
Visualization}. doi:10.32614/CRAN.package.scales
\url{https://doi.org/10.32614/CRAN.package.scales}, R package version 1.4.0,
\url{https://CRAN.R-project.org/package=scales}.
}
}
\seealso{
Other preprocessing: 
\code{\link{.BioTokenizeR_compute_bio_score}()},
\code{\link{.BioTokenizeR_preprocess_AA}()},
\code{\link{.BioTokenizeR_preprocess_DNA}()},
\code{\link{.BioTokenizeR_preprocess_RNA}()},
\code{\link{preprocess_sequences}()}
}
\concept{preprocessing}
\keyword{internal}
\keyword{preprocessing}
