% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_data}
\alias{generate_data}
\title{Generate Sequencing Data}
\usage{
generate_data(
  n = 1,
  length = 21,
  vocab_size = 10,
  preprocess = TRUE,
  annotate = TRUE,
  tokenize = TRUE,
  summarize = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{n}{A numerical value indicating the number of sequences to generate.}

\item{length}{A numerical value indicating the length of all sequences.}

\item{vocab_size}{A numerical value indicating the BPE tokenization
vocabulary size for use by \code{tokenize_sequences()}}

\item{preprocess}{A Boolean value toggling the preprocessing step.}

\item{annotate}{A Boolean value toggling the annotation step.}

\item{tokenize}{A Boolean value toggling the tokenization step.}

\item{summarize}{A Boolean value toggling the analysis step.}

\item{verbose}{A Boolean value toggling verbose output during generation.}
}
\value{
A list containing:
\describe{
\item{\code{dna_seq}}{A \code{Biostrings::DNAStringSet} of generated sequences.}
\item{\code{rna_seq}}{A \code{Biostrings::RNAStringSet} of generated sequences.}
\item{\code{aa_seq}}{A \code{Biostrings::AAStringSet} of generated sequences.}
\item{\code{dna_letters}}{A string of all letters from \code{dna_seq}.}
\item{\code{rna_letters}}{A string of all letters from \code{rna_seq}.}
\item{\code{aa_letters}}{A string of all letters from \code{ana_seq}.}
\item{\code{dna_preproc}}{A \code{bioBPE_preprocessed} object of preprocessed
DNA sequences.}
\item{\code{rna_preproc}}{A \code{bioBPE_preprocessed} object of preprocessed
RNA sequences.}
\item{\code{aa_preproc}}{A \code{bioBPE_preprocessed} object of preprocessed
AA sequences.}
\item{\code{dna_annot}}{A \code{bioBPE_preprocessed} object of preprocessed DNA
sequences with biological annotations.}
\item{\code{rna_annot}}{A \code{bioBPE_preprocessed} object of preprocessed RNA
sequences with biological annotations.}
\item{\code{aa_annot}}{A \code{bioBPE_preprocessed} object of preprocessed AA
sequences with biological annotations.}
\item{\code{dna_tokens}}{A list of tokenized DNA sequences}
\item{\code{rna_tokens}}{A list of tokenized RNA sequences}
\item{\code{aa_tokens}}{A list of tokenized AA sequences}
\item{\code{dna_summary}}{A \code{bioBPE_summary} object summarizing DNA tokens.}
\item{\code{rna_summary}}{A \code{bioBPE_summary} object summarizing RNA tokens.}
\item{\code{aa_summary}}{A \code{bioBPE_summary} object summarizing AA tokens.}
}
}
\description{
Wraps the generation of DNA, RNA, and AA sequences at different states of the
pipeline. Parameters and pipeline steps are toggled. For use by
documentation, testthat, and various tests.
}
\examples{
\dontrun{
   # Generate simulated data
   data <- generate_data(
       n          = 3, 
       length     = 1000, 
       vocab_size = 25, 
       preprocess = FALSE,
       annotate   = FALSE,
       tokenize   = FALSE,
       summarize  = FALSE,
       verbose    = FALSE
   )
}

}
\references{
{
Ponty Y, Termier M, Denise A (2006). GenRGenS: software for generating
random genomic sequences and structures. Bioinformatics, 22(12): 1534-
1535. doi:10.1093/bioinformatics/btl113.
}
}
\seealso{
Other utilities: 
\code{\link{generate_dummy_data}()},
\code{\link{generate_sequences}()}
}
\concept{utilities}
\keyword{internal}
