% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_dummy_data}
\alias{generate_dummy_data}
\title{Generate Dummy Data}
\usage{
generate_dummy_data()
}
\value{
A list containing:
\describe{
\item{\code{dna_empty}}{An empty \code{bioBPE_preprocessed} object.}
\item{\code{rna_empty}}{An empty \code{bioBPE_preprocessed} object.}
\item{\code{aa_empty}}{An empty \code{bioBPE_preprocessed} object.}
\item{\code{invalid_type}}{A \code{bioBPE_preprocessed} object of invalid type.}
}
}
\description{
Wraps the generation of the primary types of invalid sequencing data inputted
to the workflow for standardized error checking.
}
\examples{
\dontrun{
   # Generate simulated dummy data
   dummy_data <- generate_dummy_data()
}

}
\keyword{internal}
