% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_sequences}
\alias{generate_sequences}
\title{Generate Biological Sequences}
\usage{
generate_sequences(type = NULL, n = 1, length = 21)
}
\value{
A list containing:
\describe{
\item{\code{dna_empty}}{An empty \code{bioBPE_preprocessed} object.}
\item{\code{rna_empty}}{An empty \code{bioBPE_preprocessed} object.}
\item{\code{aa_empty}}{An empty \code{bioBPE_preprocessed} object.}
\item{\code{invalid_type}}{A \code{bioBPE_preprocessed} object of invalid type.}
}
}
\description{
Wraps the generation of the primary types of invalid sequencing data inputted
to the workflow for standardized error checking.
}
\examples{
\dontrun{
   # Generate simulated dummy data
   dummy_data <- generate_dummy_data()
}

}
\references{
{
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2025). Biostrings:
Efficient manipulation of biological strings.
doi:10.18129/B9.bioc.Biostrings https://doi.org/10.18129/B9.bioc.Biostrings,
R package version 2.78.0, https://bioconductor.org/packages/Biostrings.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Ponty Y, Termier M, Denise A (2006). GenRGenS: software for generating
random genomic sequences and structures. Bioinformatics, 22(12): 1534-
1535. doi:10.1093/bioinformatics/btl113.
}\if{html}{\out{</div>}}

}
}
\seealso{
Other utilities: 
\code{\link{generate_data}()},
\code{\link{generate_dummy_data}()}
}
\concept{utilities}
\keyword{internal}
