% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BioTokenizeR.R
\name{run_BioTokenizeR}
\alias{run_BioTokenizeR}
\title{Launch Shiny App for BioTokenizeR}
\usage{
run_BioTokenizeR()
}
\description{
This function launches the Shiny app for the BioTokenizeR package, displaying
an interactive web app user interface.
}
\examples{
\dontrun{
    BioTokenizeR::run_BioTokenizeR()
}

}
