% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_tokens.R
\name{plot_cumulative_coverage}
\alias{plot_cumulative_coverage}
\title{Plot Cumulative Token Frequency Coverage}
\usage{
plot_cumulative_coverage(statistics, output_dir = NULL)
}
\arguments{
\item{statistics}{An object of class \code{"bioBPE_summary"} containing token
frequency data.}

\item{output_dir}{Optional directory to save the plot. If \code{NULL}, the plot
is returned but not saved.}
}
\value{
A \code{ggplot2} object representing cumulative token frequency coverage.
}
\description{
Generates a line plot showing the cumulative frequency coverage of tokens
in the tokenized sequences, ranked from most to least frequent.
}
\examples{
\dontrun{
   # Generate simulated data
   data <- generate_data(
       n          = 3, 
       length     = 1000, 
       vocab_size = 25, 
       preprocess = TRUE,
       annotate   = TRUE,
       tokenize   = TRUE,
       summarize  = TRUE,
       verbose    = FALSE
   )
   
   # Visualize the cumulative coverage of all tokens
   plot <- plot_cumulative_coverage(statistics = data$dna_summary,
                                    output_dir = NULL)
}

}
\references{
{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Wickham H, François R, Henry L, Müller K, Vaughan D (2023). _dplyr: A 
Grammar of Data Manipulation_. doi:10.32614/CRAN.package.dplyr 
<https://doi.org/10.32614/CRAN.package.dplyr>, R package version 1.1.4,
<https://CRAN.R-project.org/package=dplyr>.
}\if{html}{\out{</div>}}

}
}
\seealso{
Other visualization: 
\code{\link{plot_token_frequency_distribution}()},
\code{\link{plot_top_tokens}()},
\code{\link{visualize_tokens}()}
}
\concept{visualization}
\keyword{plotting}
\keyword{tokenization}
\keyword{visualization}
